+++ db/access.php
@@ -60,6 +60,14 @@ $capabilities = array(
         )
     ),
 
+    'mod/customcert:receiveissue' => array(
+        'captype' => 'read',
+        'contextlevel' => CONTEXT_MODULE,
+        'archetypes' => array(
+            'student' => CAP_ALLOW
+        )
+    ),
+
     'mod/customcert:viewreport' => array(
 
         'captype' => 'read',